requirements.txt
asgiref==3.7.2
boto3==1.36.22
botocore==1.36.22
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.5
django-cors-headers==4.2.0
django-oauth-toolkit==3.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
ecdsa==0.19.0
firebase-admin==6.2.0
idna==3.10
jmespath==1.0.1
jwcrypto==1.5.6
oauthlib==3.2.2
psycopg2==2.9.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.12.2
urllib3==2.0.4
wheel==0.45.1
wrapt==1.15.0